Add inspection-batch-with-failure workflow example
This workflow submits a Job which creates `n` pod replicas and runs them in a sequence
The pod which is being ran is faulty and fails. The workflow is supposed to wait untill
a certain amount of job failures appear and fail as well.

successCondition: `n` pods have finished successfully and are present in the cluster
failureCondition: there are more failures than is allowed (a constant to account for potential flakes)

The workflow simulates the scenario when we want to create a batch of inspections but the stack is faulty
and we don't want to pollute the namespace unnecessarily with failed pods.

Signed-off-by: Marek Cermak <macermak@redhat.com>

new file:   examples/inspection-batch-with-failure.yaml